refactor: moving monopoly modules into directory
