+++ ascii/moneybag
@@ -1,14 +1,19 @@
-⠀⠀⠀⠀⠀⠀⢯⠙⠩⠙⡟⠊⠽⢖⠆⠀⠀⠀⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠱⣠⠀⢁⣄⠔⠁⠀⠀⠀⠀⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⣷⣶⣾⣾⠀⠀⠀⠀⠀⠀⠀⠀
-⠀⠀⠀⠀⠀⠀⢀⡔⠙⠈⢱⡟⣧⠀⠀⠀⠀⠀⠀⠀
-⠀⠀⠀⠀⠀⡠⠊⠀⠀⣀⡀⠀⠘⠕⢄⠀⠀⠀⠀⠀
-⠀⠀⠀⢀⠞⠀⠀⢀⣠⣿⣧⣀⠀⠀⢄⠱⡀⠀⠀⠀
-⠀⠀⡰⠃⠀⠀⢠⣿⠿⣿⡟⢿⣷⡄⠀⠑⢜⢆⠀⠀
-⠀⢰⠁⠀⠀⠀⠸⣿⣦⣿⡇⠀⠛⠋⠀⠨⡐⢍⢆⠀
-⠀⡇⠀⠀⠀⠀⠀⠙⠻⣿⣿⣿⣦⡀⠀⢀⠨⡒⠙⡄
-⢠⠁⡀⠀⠀⠀⣤⡀⠀⣿⡇⢈⣿⡷⠀⠠⢕⠢⠁⡇
-⠸⠀⡕⠀⠀⠀⢻⣿⣶⣿⣷⣾⡿⠁⠀⠨⣐⠨⢀⠃
-⠀⠣⣩⠘⠀⠀⠀⠈⠙⣿⡏⠁⠀⢀⠠⢁⡂⢉⠎⠀
+                              ⢀⡔⠙⠩⠙⡟⠊⠽⢖⠆
+                               ⠱⣠⠀⢁⣄⠔⠁
+                               ⣀⣶⣾⣾⠀⠀⠀⠀
+                             ⡔⠙⠈⢱⡟⣧⠀
+                          ⡠⠊⠀⠀⣀⡀⠀⠘⠕⢄
+⠀⠀⠀⠀⠀⠀⢯⠙⠩⠙⡟⠊⠽⢖⠆⠀⠀⠀⠀⠀⠀⠀⠀⢀⠞⠀⠀⢀⣠⣿⣧⣀⠀⠀⢄⠱⡀⠀⠀⠀
+⠀⠀⠀⠀⠀⠀⠀⠱⣠⠀⢁⣄⠔⠁⠀⠀⠀⠀⠀⠀⠀⠀⡰⠃⠀⠀⢠⣿⠿⣿⡟⢿⣷⡄⠀⠑⢜⢆⠀⠀
+⠀⠀⠀⠀⠀⠀⠀⠀⣷⣶⣾⣾⠀⠀⠀⠀⠀⠀⠀⠀⠀⢰⠁⠀⠀⠀⠸⣿⣦⣿⡇⠀⠛⠋⠀⠨⡐⢍⢆⠀
+⠀⠀⠀⠀⠀⠀⢀⡔⠙⠈⢱⡟⣧⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠙⠻⣿⣿⣿⣦⡀⠀⢀⠨⡒⠙⡄
+⠀⠀⠀⠀⠀⡠⠊⠀⠀⣀⡀⠀⠘⠕⢄⠀⠀⠀⠀⠀⢠⠁⡀⠀⠀⠀⣤⡀⠀⣿⡇⢈⣿⡷⠀⠠⢕⠢⠁⡇
+⠀⠀⠀⢀⠞⣦⠀⢀⣠⣿⣧⣀⠀⠀⢄⠱⡀⠀⠀⠀⠸ ⡕⠀⠀⠀⢻⣿⣶⣿⣷⣾⡿⠁⠀⠨⣐⠨⢀⠃
+⠀⠀⡰⠃⡿⠀⢠⣿⠿⣿⡟⢿⣷⡄⠀⠑⢜⢆⠀⠀⠀⠣⣩⠘⠀⠀⠀⠈⠙⣿⡏⠁⠀⢀⠠⢁⡂⢉⠎
+⠀⢰⠁⡏⣦⠀⠸⣿⣦⣿⡇⠀⠛⠋⠀  ⡐⢍⢆⠀⠀⠀⠈⠓⠬⢀⣀⠀⠀⠈⠀⠀⠀⢐⣬⠴⠒⠁
+⠀⡇⠨⠀⠀ ⠀⠙⠻⣿⣿⣿⣦⡀⠀⢀⠨  ⠙⡄⠀⠀⠀⠀⠀⠀⠀⠈⠉⠉⠉⠉⠉⠀
+⢠⠁⡀⠨⠀⠀⣤⡀⠀⣿⡇⢈⣿⡷⠀   ⠢⠁⡇
+⠸⠀⡕⣦⠀⠀⢻⣿⣶⣿⣷⣾⡿⠁⠀⠨⣐ ⢀⠃
+⠀⠣⣩⠘⠀⠀⠀⠈⠙⣿⡏⠁⠀⢀⠠ ⡂⢉⠎⠀
 ⠀⠀⠈⠓⠬⢀⣀⠀⠀⠈⠀⠀⠀⢐⣬⠴⠒⠁⠀⠀
 ⠀⠀⠀⠀⠀⠀⠀⠈⠉⠉⠉⠉⠉⠀⠀⠀⠀⠀⠀⠀
