--- errorlog.txt
@@ -1,24 +1,0 @@
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
-Player -1 already in game.Line178
