+++ monopoly_directory/properties.py
