+++ screenspace.py
@@ -2,8 +2,8 @@
 
 # Player Terminal total width and height: 153x43. There is 3 extra characters for the border.
 # Banker total width and height is 200x60
-WIDTH = 150
-HEIGHT = 40
+WIDTH = 200
+HEIGHT = 60
 INPUTLINE = 45
 import os
 from style import MYCOLORS as COLORS, choose_colorset, set_cursor, set_cursor_str, colortest, graphics as g
